Updated README and setup script.
Changed file name of README.md to README.rst to reflect the content
of the file which is formatted as restructured text. Added a short
intro section, an example and updated the installatian guide.

setup.py now reads the README file and uses it for the long
description for the package. That is also the text that is
shown on PyPI. Furthermore the URL was updated to the new github
page under the laspy organization and the development status was
changed to 'Production/Stable'.