Accept non-ascii string when reading.
This commit makes it so that for the header's `generating_software` and `system_identifier`
as well as the vlr's `description' laspy will now handle decoding errors by storing the raw bytes instead of
the decoded string when such an error happens.

However when writing, the correct encoding will still be enforced.

That way we are able to read files with invalid encodings but we won't write files with invalid encoding.