Fix header.update mechanism
The `update` function of the `LasHeader` was written in a way
that made it compute the correct x,y,z bound when writing in chunk mode.

`update` computed the bounds by "growing" them, which is the right strategy
for chunk mode writing, but not when you juste want to update the header
to match the values  after slicing or setting the `LasData.points`.

We fix this by adding a `grow` method to the header and use it in appropriate places
and the `update` is changed to actually match the set of point given.