Fix desync between LasData.header.scales and LasData.points.scales
This is a second regression introduced by 8426cbc (so its a bug that does not appear in 2.0 but it does in 2.1).

Since 8426cbc , the scales and offsets are also stored in the internal points array of the LasData,
and changing the scales and/or offsets by doing las.header.scales = would not updated the offsets and scales stored in
the points array making them out of sync.
So when setting the x,y or z coordinates after that would not produce the expected result.