Move evlrs into LasHeader class
The EVLRs are now stored in the `LasHeader` class,
attributes are added in `LasData` and  `LasReader` for backward compatibility.

Motivation for this change is that although the header + vlrs are at the begining and the evlrs are potentially storing the same kind of vlr like the CRS one which we may want to have access to when simply opening.