Add option to allow users to skip reading evlrs when opening a file
Since 2.3.0, EVLRs are read when the file is opened by `LasReader` / `laspy.open`.

However this is not always ideal, when opening a remote file
(eg AWS S3, HTTP server without doing range requests) it means
we will seek to evlrs which are the end of the file, meaning we
will transfer the whole file.

This adds a `read_evlrs` option (True bu default) to `LasReader.__init__` and `laspy.open` to
allow users to skip that phase during initialization