Fix LasAppender when input file's last chunk is complete
When the input file had all of its chunks complete (ie all chunks contain chunk_size points)
we would incorrectly discard the last one.

We initialized the LasAppender's chunk table to the one from the input file minus the last chunk
however this is not exactly what we need.
What we need is the chunk table to be contain all the number of complete chunk.