Handle redundant CSs in GeoKeyDirectoryVlr
Some LAS files with their positioning data stored in a Projected
Coordinate System might have the corresponding Geographic Coordinate
System redundantly set. The current parse_crs implementation will spit
the first CRS it finds in the GeoKeyDirectoryVlr, yielding the wrong one
when both are set.

This PR will solve this issue by prioritizing the Projected CS when both
are set. The autzen.las file was updated to include this redundancy in
its headers.