Close input/output stream if on opening fail
The `laspy.open` (open_las) was meant to be used with a
`with statement` (`with laspy.open(....) as f:...`) so that the input/ouput
would properly be closed in case of  and exception being raised.

However, if the exception got raised within the
`__init__` function of either of LasReader, LasWriter, LasAppender
then the object in `as name` would not have been constructed thus
the `__exit__` function of the LasReader/LasWriter/LasAppender
cannot be run, meaning the stream would not be closed.

This commit fixes that by catching, closing file then rethrowing in
the open_las function