Check  geokey value before assuming its an EPSG
In the geotiff specification, the value (contained in value_offset)
for the keys ProjectedCRSGeoKey and GeodeticCRSGeoKey had different
meaning depending on the range in which the value was.

We however assumed the value was always an EPSG code,
leading to errors because we would then ask pyproj to
get the proj from the epsg value, which was not one.

This commit fixes it by adding a check on the range
written in the spec.