Fix: don't reset custom creation_date
LasHeader.partial_reset used to always reset the
creation_date to date.today(),
meaning users could not set a custom date.

This makes it so that if the header already has a date
where becayse manually set by the user or because it
is the value read from a file, it won't get reset

Writing a file from a newly created header
will still write date.today to the file on disk.