Automatically rescale points in writer
When the points given to the LasWriter.write_points
are of type ScaleAwarePointRecord we now automatically rescale
the points.
This is not possible for PackedPointRecord because they
do not know the scales and offsets they originate from.

The goal is to make points coming from
different files (with possibly different scales/offsets)
have the same (and correct) scales/offsets, making merging
easier.